Send a shutdown signal to the sparkle task when the system transistions from sparkle mode to play mode.
This is needed because of a race condition that can be encountered if
the user presses PLAY while the system is in the middle of calling
sparkle_task->display_update_leds() and sending data to the led1642gw
drivers.

If we abruptly kill the sparkle_task, the led1642gw_lock could remain
active and will never clear, causing the system to deadlock as soon as
the play_game() function starts.