Ignore subsequent GPIO-high events that fire when the play button is released.
For some reason when the play button is strung and hidden in a christmas
tree, releasing the button triggers multiple GPIO-high events.  This
anomaly isn't noticable when playing the game, but it makes precise
theme selection impossible.

Also ensure the timer is stopped after sucessfully firing once.  We
don't want to cycle in and out of theme-select mode by holding down the
button for multiple seconds.